+++ example/lib/screens/StyleSettings.dart
@@ -1,5 +1,4 @@
 import '../utils/utils.dart';
-import '../l10n/app_localizations.dart';
 
 import 'package:empathetech_flutter_ui/empathetech_flutter_ui.dart';
 
@@ -19,18 +18,17 @@ class _StyleSettingsScreenState extends State<StyleSettingsScreen> {
   @override
   void initState() {
     super.initState();
-    setPageTitle(
-        context: context, title: AppLocalizations.of(context)!.styleSettings);
+    setPageTitle(context: context, title: Phrases.of(context)!.styleSettings);
   }
 
   // Gather theme data //
 
   final double _buttonSpacer = EzConfig.instance.prefs[buttonSpacingKey];
 
-  late final String _resetTitle = AppLocalizations.of(context)!.resetAllStyle;
+  late final String _resetTitle = Phrases.of(context)!.resetAllStyle;
   late final String _resetMessage = kIsWeb
-      ? AppLocalizations.of(context)!.resetAllWarningWeb
-      : AppLocalizations.of(context)!.resetAllWarning;
+      ? Phrases.of(context)!.resetAllWarningWeb
+      : Phrases.of(context)!.resetAllWarning;
 
   // Return the build //