+++ CHANGELOG.md
@@ -7,7 +7,7 @@ All notable changes to this project will be documented in this file.
 - Create git release section of Jenkinsfile
 - Reverted line length
 - Fixed README links
-- EFUIPhrases to EFUIPhrases and AppLocalizations to Phrases
+- EFUILocalizations to EFUIPhrases and AppLocalizations to Phrases
   - Minor breaking change
 
 ## [4.0.0] - 2023-10-07